moving payload work to Core
