+++ dotnet-build.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Faction.Common" Version="2019.10.3-beta" />
+    <!-- <PackageReference Include="Faction.Common" Version="2019.10.3-beta" /> -->
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.0" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.0" />
@@ -20,8 +20,11 @@
     <PackageReference Include="RabbitMQ" Version="3.6.2" />
     <PackageReference Include="RabbitMQ.Client" Version="5.1.2" />
     <!-- This is used for testing new versions of the common library -->
-    <!-- <Reference Include="Faction.Common">
-      <HintPath>.\lib\Faction.Common.dll</HintPath>
-    </Reference> -->
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="Faction.Common, Version=2019.10.3.0, Culture=neutral, PublicKeyToken=null">
+      <HintPath>..\Faction.Common\bin\Debug\netstandard2.1\Faction.Common.dll</HintPath>
+    </Reference>
   </ItemGroup>
 </Project>
